Checks if plagiarized votes >= 2/3 of (plagiarized + not_plagiarized).
If yes, sets status of member = 'blacklisted'. When that happens, all their texts are also going to be not displayed anymore, we'll set their texts to 'archiv'. Must make sure of that in the views of list of texts in a table format (like in item.php or in Browse).